Refactored ApplicationViews, App, main, and index
