+++ src/main.tsx
@@ -1,0 +1,15 @@
+import ReactDOM from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
+import { App } from "./App.js";
+import "./index.css";
+
+const container = document.getElementById("root");
+
+if (container) {
+  const root = ReactDOM.createRoot(container);
+  root.render(
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
+  );
+}
