+++ src/views/Authorized.tsx
@@ -1,9 +1,14 @@
+import { ReactNode } from "react";
 import { Navigate, useLocation } from "react-router-dom";
+
+interface AuthorizedProps {
+  children: ReactNode;
+}
 
 // We can access child components the same way we access props. Child components are passed to our props as a key/value pair where
 // children is the key.
 
-export const Authorized = ({ children }) => {
+export const Authorized = ({ children }: AuthorizedProps) => {
   let location = useLocation();
 
   // Check if user is logged in. If they are, render the CHILD components (in this case, the ApplicationViews component)
